requirements.txt
# ===========================================
# Performia System - Complete Requirements
# ===========================================

# Core System Dependencies
pyautogen>=0.2.0
numpy>=1.24.0
python-osc>=1.8.0
pyyaml>=6.0

# Audio & Synthesis
# supercollider>=0.4.0  # Removed - install via brew/apt instead
sounddevice>=0.4.6      # Audio input/output
librosa>=0.10.0         # Audio analysis
scipy>=1.10.0           # Signal processing
soundfile>=0.12.1       # Audio file I/O

# MIDI Control
mido>=1.3.0             # MIDI messages
python-rtmidi>=1.5.0    # MIDI backend

# Performance Optimization
uvloop>=0.17.0 ; platform_system != "Windows"
numba>=0.57.0           # JIT compilation

# GUI Backend
flask>=3.0.0            # Web server
flask-cors>=4.0.0       # CORS support
flask-socketio>=5.3.0   # WebSocket support
python-socketio>=5.10.0 # Socket.IO server
eventlet>=0.33.0        # Async networking

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0

# Monitoring & Visualization
matplotlib>=3.6.0       # Plotting
pygame>=2.1.0          # Real-time visualization
psutil>=5.9.0          # System monitoring